Update encoder to deal with different inputs
String#encode is not executed if the string has the same encoding of the
argument, even if it contains invalid chars. In those occasions the iconv
library was used to force the conversion anyway.

This remove the deprecated (removed) iconv library and use only the standard
ruby library method "force_encoding" method which does force the
conversion anyway.