Enable the use of headers when generating a HTTP request
Previously, there was no way of passing through custom headers when
using Catcher. This meant that when hitting some third-party API's which
require authentication, we would always get back a 403 Unauthorised.

This commit addresses this by allowing an inheriting class of
'Catcher::API' to specify its own 'header' method, which will contain
their required headers so that we can gain access and get back the
correct data from an API.